Return integer or float in captures
In order to save us from calling `String.to_integer/1` or
`String.to_float/1` in our step definition for values which are integers
or floats.

So far it's a quick `check the concept` implementation, which is a good
point to stop for a discussion. Once we agreed that this feature make
sense we can go and tidy it up.

Note: this feature is only available for Cucumber Expressions, as we
want to be backwards compatible and don't introduce breaking changes :)